Base backend done. Missing checking for correctly use and implementation of JWT token, and improving the post/put methods.

+++ src/main/kotlin/com/dlp/dlp_api/entity/CustomPattern.kt
@@ -14,7 +14,7 @@ class CustomPattern (
     @Column(nullable = false)
     var regex: String,
     @Column(nullable = false)
-    var score: Int,
+    var score: Float,
 
     @ManyToOne
     @JoinColumn(name="entity_type_id")
